requirements.txt
Django==1.9.1
mysqlclient
djangorestframework
markdown
django-filter
django-cors-headers
celery
django-celery
scrapy
requests==2.9.1
